Ejercicio 6 de la unidad 2 resuelto (corrección).

+++ src/unidad2/Colores.java
@@ -37,7 +37,7 @@ public class Colores {
 						break;
 				}
 				
-				int componenteRGB = (int) Long.parseLong(in.readLine());
+				int componenteRGB = Integer.parseInt(in.readLine());
 				rgb[i][j] = componenteRGB;
 			}
